refactor: deduplicate coverage and test runner plugins to fix duplication threshold
Extract shared logic into base classes to reduce code duplication from 6.3%
to well below the 5.0% threshold:

- CoveragePlugin base: shared Istanbul report parsing, issue creation, no-data issues
- TestRunnerPlugin base: shared Jest-compatible report processing, assertion handling
- Remove duplicate _create_no_data_issue from coverage_py, jacoco, tarpaulin
- Remove duplicate _create_coverage_issue from coverage_py, jacoco
