+++ tests/unit/plugins/test_runners/test_jest.py
@@ -18,17 +18,14 @@ class TestJestRunner:
     """Tests for JestRunner class."""
 
     def test_name(self) -> None:
-        """Test plugin name."""
         runner = JestRunner()
         assert runner.name == "jest"
 
     def test_languages(self) -> None:
-        """Test supported languages."""
         runner = JestRunner()
         assert runner.languages == ["javascript", "typescript"]
 
     def test_domain(self) -> None:
-        """Test domain is TESTING."""
         runner = JestRunner()
         assert runner.domain == ToolDomain.TESTING
 
@@ -37,7 +34,6 @@ class TestJestRunnerBinaryFinding:
     """Tests for binary finding logic."""
 
     def test_find_in_node_modules(self) -> None:
-        """Test finding jest in project node_modules."""
         with tempfile.TemporaryDirectory() as tmpdir:
             project_root = Path(tmpdir)
             node_bin = project_root / "node_modules" / ".bin"
@@ -48,28 +44,21 @@ class TestJestRunnerBinaryFinding:
 
             runner = JestRunner(project_root=project_root)
             binary = runner.ensure_binary()
-
             assert binary == jest_bin
 
     @patch("shutil.which")
     def test_find_in_system_path(self, mock_which: MagicMock) -> None:
-        """Test finding jest in system PATH."""
         mock_which.return_value = "/usr/local/bin/jest"
-
         runner = JestRunner()
         binary = runner.ensure_binary()
-
         assert binary == Path("/usr/local/bin/jest")
 
     @patch("shutil.which")
     def test_not_found_raises_error(self, mock_which: MagicMock) -> None:
-        """Test FileNotFoundError when jest not found."""
         mock_which.return_value = None
-
         runner = JestRunner()
         with pytest.raises(FileNotFoundError) as exc:
             runner.ensure_binary()
-
         assert "Jest is not installed" in str(exc.value)
 
 
@@ -77,7 +66,6 @@ class TestJestGetVersion:
     """Tests for version detection."""
 
     def test_get_version_success(self) -> None:
-        """Test getting Jest version."""
         with tempfile.TemporaryDirectory() as tmpdir:
             project_root = Path(tmpdir)
             node_bin = project_root / "node_modules" / ".bin"
@@ -87,14 +75,12 @@ class TestJestGetVersion:
             jest_bin.chmod(0o755)
 
             runner = JestRunner(project_root=project_root)
-
             with patch("lucidshark.plugins.test_runners.jest.get_cli_version", return_value="29.7.0"):
                 version = runner.get_version()
                 assert version == "29.7.0"
 
     @patch("shutil.which", return_value=None)
     def test_get_version_unknown_when_not_found(self, mock_which: MagicMock) -> None:
-        """Test version returns 'unknown' when jest not found."""
         runner = JestRunner()
         version = runner.get_version()
         assert version == "unknown"
@@ -105,7 +91,6 @@ class TestJestRunTests:
 
     @patch("shutil.which", return_value=None)
     def test_run_tests_binary_not_found(self, mock_which: MagicMock) -> None:
-        """Test run_tests returns empty result when binary not found."""
         runner = JestRunner()
         context = MagicMock()
         context.project_root = Path("/project")
@@ -117,7 +102,6 @@ class TestJestRunTests:
         assert result.failed == 0
 
     def test_run_tests_timeout(self) -> None:
-        """Test run_tests handles timeout."""
         with tempfile.TemporaryDirectory() as tmpdir:
             project_root = Path(tmpdir)
             node_bin = project_root / "node_modules" / ".bin"
@@ -136,7 +120,6 @@ class TestJestRunTests:
                 assert result.passed == 0
 
     def test_run_tests_general_exception(self) -> None:
-        """Test run_tests handles general exceptions."""
         with tempfile.TemporaryDirectory() as tmpdir:
             project_root = Path(tmpdir)
             node_bin = project_root / "node_modules" / ".bin"
@@ -155,7 +138,6 @@ class TestJestRunTests:
                 assert result.passed == 0
 
     def test_run_tests_always_includes_coverage_flag(self) -> None:
-        """Test that --coverage flag is always added."""
         with tempfile.TemporaryDirectory() as tmpdir:
             project_root = Path(tmpdir)
             node_bin = project_root / "node_modules" / ".bin"
@@ -178,7 +160,6 @@ class TestJestRunTests:
                 assert "--coverage" in cmd
 
     def test_run_tests_with_paths(self) -> None:
-        """Test that paths are appended to command."""
         with tempfile.TemporaryDirectory() as tmpdir:
             project_root = Path(tmpdir)
             node_bin = project_root / "node_modules" / ".bin"
@@ -201,7 +182,6 @@ class TestJestRunTests:
                 assert "src/tests" in cmd
 
     def test_run_tests_parses_json_report_from_file(self) -> None:
-        """Test parsing JSON report from output file."""
         with tempfile.TemporaryDirectory() as tmpdir:
             project_root = Path(tmpdir)
             node_bin = project_root / "node_modules" / ".bin"
@@ -215,9 +195,7 @@ class TestJestRunTests:
             context.project_root = project_root
             context.paths = []
 
-            # Mock subprocess to write a JSON report file
             def fake_run(cmd, **kwargs):
-                # Find the outputFile argument
                 for arg in cmd:
                     if arg.startswith("--outputFile="):
                         report_path = Path(arg.split("=", 1)[1])
@@ -241,10 +219,9 @@ class TestJestRunTests:
 
 
 class TestJestReportProcessing:
-    """Tests for Jest report processing."""
+    """Tests for Jest report processing (via base class)."""
 
     def test_process_report_with_failures(self) -> None:
-        """Test processing Jest report with failures."""
         runner = JestRunner()
 
         report = {
@@ -292,7 +269,6 @@ class TestJestReportProcessing:
         assert issue.source_tool == "jest"
 
     def test_process_report_all_passed(self) -> None:
-        """Test processing Jest report with all tests passed."""
         runner = JestRunner()
 
         report = {
@@ -312,7 +288,6 @@ class TestJestReportProcessing:
         assert len(result.issues) == 0
 
     def test_process_report_with_todo_tests(self) -> None:
-        """Test todo tests are counted as skipped."""
         runner = JestRunner()
 
         report = {
@@ -327,7 +302,6 @@ class TestJestReportProcessing:
         assert result.skipped == 5  # 2 pending + 3 todo
 
     def test_process_report_duration_calculation(self) -> None:
-        """Test duration is calculated from testResults."""
         runner = JestRunner()
 
         report = {
@@ -346,28 +320,24 @@ class TestJestReportProcessing:
 
 
 class TestJestJsonOutput:
-    """Tests for JSON output parsing."""
+    """Tests for JSON output parsing (via base class)."""
 
     def test_parse_json_output_empty(self) -> None:
-        """Test empty output returns empty result."""
         runner = JestRunner()
         result = runner._parse_json_output("", Path("/project"))
         assert result.passed == 0
 
     def test_parse_json_output_whitespace_only(self) -> None:
-        """Test whitespace-only output returns empty result."""
         runner = JestRunner()
         result = runner._parse_json_output("   \n  ", Path("/project"))
         assert result.passed == 0
 
     def test_parse_json_output_invalid_json(self) -> None:
-        """Test invalid JSON returns empty result."""
         runner = JestRunner()
         result = runner._parse_json_output("not json {", Path("/project"))
         assert result.passed == 0
 
     def test_parse_json_output_valid(self) -> None:
-        """Test valid JSON output is parsed correctly."""
         runner = JestRunner()
         output = json.dumps({
             "numPassedTests": 5,
@@ -380,7 +350,6 @@ class TestJestJsonOutput:
         assert result.passed == 5
 
     def test_parse_json_report_file_not_readable(self) -> None:
-        """Test handling unreadable JSON report file."""
         runner = JestRunner()
 
         with tempfile.TemporaryDirectory() as tmpdir:
@@ -392,10 +361,9 @@ class TestJestJsonOutput:
 
 
 class TestJestAssertionToIssue:
-    """Tests for assertion to issue conversion."""
+    """Tests for assertion to issue conversion (via base class)."""
 
     def test_assertion_with_location(self) -> None:
-        """Test converting assertion with location info."""
         runner = JestRunner()
 
         assertion = {
@@ -419,7 +387,6 @@ class TestJestAssertionToIssue:
         assert issue.severity == Severity.HIGH
 
     def test_assertion_without_location(self) -> None:
-        """Test converting assertion without location."""
         runner = JestRunner()
 
         assertion = {
@@ -440,7 +407,6 @@ class TestJestAssertionToIssue:
         assert issue.line_start is None
 
     def test_assertion_relative_file_path(self) -> None:
-        """Test that relative file paths are resolved against project root."""
         runner = JestRunner()
 
         assertion = {
@@ -451,7 +417,7 @@ class TestJestAssertionToIssue:
             "failureMessages": ["fail"],
         }
         test_file = {
-            "name": "tests/app.test.js",  # relative path
+            "name": "tests/app.test.js",
             "status": "failed",
         }
 
@@ -460,7 +426,6 @@ class TestJestAssertionToIssue:
         assert issue.file_path == Path("/project/tests/app.test.js")
 
     def test_assertion_no_failure_messages(self) -> None:
-        """Test handling assertion with no failure messages."""
         runner = JestRunner()
 
         assertion = {
@@ -478,91 +443,55 @@ class TestJestAssertionToIssue:
 
 
 class TestJestAssertionExtraction:
-    """Tests for assertion message extraction."""
+    """Tests for assertion message extraction (via base class)."""
 
     def test_extract_expect_pattern(self) -> None:
-        """Test extracting expect() assertion."""
         runner = JestRunner()
-
-        message = """
-expect(received).toBe(expected)
-
-Expected: 2
-Received: 1
-        """
-
+        message = "expect(received).toBe(expected)\n\nExpected: 2\nReceived: 1"
         result = runner._extract_assertion(message)
         assert "Expected:" in result or "expect" in result.lower()
 
     def test_extract_expected_received_pattern(self) -> None:
-        """Test extracting Expected/Received pattern."""
         runner = JestRunner()
-
-        message = """
-Expected: 42
-Received: 0
-        """
-
+        message = "\nExpected: 42\nReceived: 0\n"
         result = runner._extract_assertion(message)
         assert "Expected:" in result
 
     def test_extract_first_meaningful_line(self) -> None:
-        """Test extracting first meaningful line as fallback."""
         runner = JestRunner()
-
-        message = """
-TypeError: Cannot read property 'foo' of undefined
-    at Object.<anonymous> (test.js:5:1)
-        """
-
+        message = "\nTypeError: Cannot read property 'foo' of undefined\n    at Object.<anonymous> (test.js:5:1)\n"
         result = runner._extract_assertion(message)
         assert "TypeError" in result
 
     def test_empty_message(self) -> None:
-        """Test empty message returns empty string."""
         runner = JestRunner()
         result = runner._extract_assertion("")
         assert result == ""
 
     def test_extract_skips_short_lines(self) -> None:
-        """Test that short lines (<=5 chars) are skipped in fallback."""
         runner = JestRunner()
-
-        message = """
-at Object.test
-Some meaningful error message here
-        """
-
+        message = "\nat Object.test\nSome meaningful error message here\n"
         result = runner._extract_assertion(message)
         assert len(result) > 5
 
 
 class TestJestIssueIdGeneration:
-    """Tests for deterministic issue ID generation."""
+    """Tests for deterministic issue ID generation (via base class)."""
 
     def test_same_input_same_id(self) -> None:
-        """Test same input produces same ID."""
         runner = JestRunner()
-
         id1 = runner._generate_issue_id("Test > should work", "expect")
         id2 = runner._generate_issue_id("Test > should work", "expect")
-
         assert id1 == id2
 
     def test_different_input_different_id(self) -> None:
-        """Test different input produces different ID."""
         runner = JestRunner()
-
         id1 = runner._generate_issue_id("Test > should work", "expect 1")
         id2 = runner._generate_issue_id("Test > should fail", "expect 2")
-
         assert id1 != id2
 
     def test_id_format(self) -> None:
-        """Test ID format starts with jest-."""
         runner = JestRunner()
-
         issue_id = runner._generate_issue_id("Test > should work", "expect")
-
         assert issue_id.startswith("jest-")
-        assert len(issue_id) == len("jest-") + 12  # 12 char hash
+        assert len(issue_id) == len("jest-") + 12
