+++ tests/unit/plugins/test_runners/test_vitest.py
@@ -18,17 +18,14 @@ class TestVitestRunner:
     """Tests for VitestRunner class."""
 
     def test_name(self) -> None:
-        """Test plugin name."""
         runner = VitestRunner()
         assert runner.name == "vitest"
 
     def test_languages(self) -> None:
-        """Test supported languages."""
         runner = VitestRunner()
         assert runner.languages == ["javascript", "typescript"]
 
     def test_domain(self) -> None:
-        """Test domain is TESTING."""
         runner = VitestRunner()
         assert runner.domain == ToolDomain.TESTING
 
@@ -37,7 +34,6 @@ class TestVitestRunnerBinaryFinding:
     """Tests for binary finding logic."""
 
     def test_find_in_node_modules(self) -> None:
-        """Test finding vitest in project node_modules."""
         with tempfile.TemporaryDirectory() as tmpdir:
             project_root = Path(tmpdir)
             node_bin = project_root / "node_modules" / ".bin"
@@ -48,28 +44,21 @@ class TestVitestRunnerBinaryFinding:
 
             runner = VitestRunner(project_root=project_root)
             binary = runner.ensure_binary()
-
             assert binary == vitest_bin
 
     @patch("shutil.which")
     def test_find_in_system_path(self, mock_which: MagicMock) -> None:
-        """Test finding vitest in system PATH."""
         mock_which.return_value = "/usr/local/bin/vitest"
-
         runner = VitestRunner()
         binary = runner.ensure_binary()
-
         assert binary == Path("/usr/local/bin/vitest")
 
     @patch("shutil.which")
     def test_not_found_raises_error(self, mock_which: MagicMock) -> None:
-        """Test FileNotFoundError when vitest not found."""
         mock_which.return_value = None
-
         runner = VitestRunner()
         with pytest.raises(FileNotFoundError) as exc:
             runner.ensure_binary()
-
         assert "Vitest is not installed" in str(exc.value)
 
 
@@ -77,7 +66,6 @@ class TestVitestGetVersion:
     """Tests for version detection."""
 
     def test_get_version_success(self) -> None:
-        """Test getting Vitest version."""
         with tempfile.TemporaryDirectory() as tmpdir:
             project_root = Path(tmpdir)
             node_bin = project_root / "node_modules" / ".bin"
@@ -87,7 +75,6 @@ class TestVitestGetVersion:
             vitest_bin.chmod(0o755)
 
             runner = VitestRunner(project_root=project_root)
-
             with patch(
                 "lucidshark.plugins.test_runners.vitest.get_cli_version",
                 return_value="3.0.4",
@@ -97,7 +84,6 @@ class TestVitestGetVersion:
 
     @patch("shutil.which", return_value=None)
     def test_get_version_unknown_when_not_found(self, mock_which: MagicMock) -> None:
-        """Test version returns 'unknown' when vitest not found."""
         runner = VitestRunner()
         version = runner.get_version()
         assert version == "unknown"
@@ -108,7 +94,6 @@ class TestVitestRunTests:
 
     @patch("shutil.which", return_value=None)
     def test_run_tests_binary_not_found(self, mock_which: MagicMock) -> None:
-        """Test run_tests returns empty result when binary not found."""
         runner = VitestRunner()
         context = MagicMock()
         context.project_root = Path("/project")
@@ -120,7 +105,6 @@ class TestVitestRunTests:
         assert result.failed == 0
 
     def test_run_tests_timeout(self) -> None:
-        """Test run_tests handles timeout."""
         with tempfile.TemporaryDirectory() as tmpdir:
             project_root = Path(tmpdir)
             node_bin = project_root / "node_modules" / ".bin"
@@ -142,7 +126,6 @@ class TestVitestRunTests:
                 assert result.passed == 0
 
     def test_run_tests_general_exception(self) -> None:
-        """Test run_tests handles general exceptions."""
         with tempfile.TemporaryDirectory() as tmpdir:
             project_root = Path(tmpdir)
             node_bin = project_root / "node_modules" / ".bin"
@@ -161,7 +144,6 @@ class TestVitestRunTests:
                 assert result.passed == 0
 
     def test_run_tests_always_includes_coverage_flag(self) -> None:
-        """Test that --coverage flag is always added."""
         with tempfile.TemporaryDirectory() as tmpdir:
             project_root = Path(tmpdir)
             node_bin = project_root / "node_modules" / ".bin"
@@ -184,7 +166,6 @@ class TestVitestRunTests:
                 assert "--coverage" in cmd
 
     def test_run_tests_with_paths(self) -> None:
-        """Test that paths are appended to command."""
         with tempfile.TemporaryDirectory() as tmpdir:
             project_root = Path(tmpdir)
             node_bin = project_root / "node_modules" / ".bin"
@@ -207,7 +188,6 @@ class TestVitestRunTests:
                 assert "src/tests" in cmd
 
     def test_run_tests_uses_run_flag(self) -> None:
-        """Test that 'run' flag is used for non-watch mode."""
         with tempfile.TemporaryDirectory() as tmpdir:
             project_root = Path(tmpdir)
             node_bin = project_root / "node_modules" / ".bin"
@@ -230,7 +210,6 @@ class TestVitestRunTests:
                 assert "run" in cmd
 
     def test_run_tests_uses_json_reporter(self) -> None:
-        """Test that --reporter=json is used."""
         with tempfile.TemporaryDirectory() as tmpdir:
             project_root = Path(tmpdir)
             node_bin = project_root / "node_modules" / ".bin"
@@ -253,7 +232,6 @@ class TestVitestRunTests:
                 assert "--reporter=json" in cmd
 
     def test_run_tests_parses_json_report_from_file(self) -> None:
-        """Test parsing JSON report from output file."""
         with tempfile.TemporaryDirectory() as tmpdir:
             project_root = Path(tmpdir)
             node_bin = project_root / "node_modules" / ".bin"
@@ -267,9 +245,7 @@ class TestVitestRunTests:
             context.project_root = project_root
             context.paths = []
 
-            # Mock subprocess to write a JSON report file
             def fake_run(cmd, **kwargs):
-                # Find the outputFile argument
                 for arg in cmd:
                     if arg.startswith("--outputFile="):
                         report_path = Path(arg.split("=", 1)[1])
@@ -293,10 +269,9 @@ class TestVitestRunTests:
 
 
 class TestVitestReportProcessing:
-    """Tests for Vitest report processing."""
+    """Tests for Vitest report processing (via base class)."""
 
     def test_process_report_with_failures(self) -> None:
-        """Test processing Vitest report with failures."""
         runner = VitestRunner()
 
         report = {
@@ -344,7 +319,6 @@ class TestVitestReportProcessing:
         assert issue.source_tool == "vitest"
 
     def test_process_report_all_passed(self) -> None:
-        """Test processing Vitest report with all tests passed."""
         runner = VitestRunner()
 
         report = {
@@ -364,7 +338,6 @@ class TestVitestReportProcessing:
         assert len(result.issues) == 0
 
     def test_process_report_with_todo_tests(self) -> None:
-        """Test todo tests are counted as skipped."""
         runner = VitestRunner()
 
         report = {
@@ -379,7 +352,6 @@ class TestVitestReportProcessing:
         assert result.skipped == 5  # 2 pending + 3 todo
 
     def test_process_report_duration_calculation(self) -> None:
-        """Test duration is calculated from testResults."""
         runner = VitestRunner()
 
         report = {
@@ -410,28 +382,24 @@ class TestVitestReportProcessing:
 
 
 class TestVitestJsonOutput:
-    """Tests for JSON output parsing."""
+    """Tests for JSON output parsing (via base class)."""
 
     def test_parse_json_output_empty(self) -> None:
-        """Test empty output returns empty result."""
         runner = VitestRunner()
         result = runner._parse_json_output("", Path("/project"))
         assert result.passed == 0
 
     def test_parse_json_output_whitespace_only(self) -> None:
-        """Test whitespace-only output returns empty result."""
         runner = VitestRunner()
         result = runner._parse_json_output("   \n  ", Path("/project"))
         assert result.passed == 0
 
     def test_parse_json_output_invalid_json(self) -> None:
-        """Test invalid JSON returns empty result."""
         runner = VitestRunner()
         result = runner._parse_json_output("not json {", Path("/project"))
         assert result.passed == 0
 
     def test_parse_json_output_valid(self) -> None:
-        """Test valid JSON output is parsed correctly."""
         runner = VitestRunner()
         output = json.dumps(
             {
@@ -446,7 +414,6 @@ class TestVitestJsonOutput:
         assert result.passed == 5
 
     def test_parse_json_report_file_not_readable(self) -> None:
-        """Test handling unreadable JSON report file."""
         runner = VitestRunner()
 
         with tempfile.TemporaryDirectory() as tmpdir:
@@ -458,10 +425,9 @@ class TestVitestJsonOutput:
 
 
 class TestVitestAssertionToIssue:
-    """Tests for assertion to issue conversion."""
+    """Tests for assertion to issue conversion (via base class)."""
 
     def test_assertion_with_location(self) -> None:
-        """Test converting assertion with location info."""
         runner = VitestRunner()
 
         assertion = {
@@ -485,7 +451,6 @@ class TestVitestAssertionToIssue:
         assert issue.severity == Severity.HIGH
 
     def test_assertion_without_location(self) -> None:
-        """Test converting assertion without location."""
         runner = VitestRunner()
 
         assertion = {
@@ -506,7 +471,6 @@ class TestVitestAssertionToIssue:
         assert issue.line_start is None
 
     def test_assertion_relative_file_path(self) -> None:
-        """Test that relative file paths are resolved against project root."""
         runner = VitestRunner()
 
         assertion = {
@@ -517,7 +481,7 @@ class TestVitestAssertionToIssue:
             "failureMessages": ["fail"],
         }
         test_file = {
-            "name": "tests/app.test.ts",  # relative path
+            "name": "tests/app.test.ts",
             "status": "failed",
         }
 
@@ -526,7 +490,6 @@ class TestVitestAssertionToIssue:
         assert issue.file_path == Path("/project/tests/app.test.ts")
 
     def test_assertion_no_failure_messages(self) -> None:
-        """Test handling assertion with no failure messages."""
         runner = VitestRunner()
 
         assertion = {
@@ -544,91 +507,55 @@ class TestVitestAssertionToIssue:
 
 
 class TestVitestAssertionExtraction:
-    """Tests for assertion message extraction."""
+    """Tests for assertion message extraction (via base class)."""
 
     def test_extract_expect_pattern(self) -> None:
-        """Test extracting expect() assertion."""
         runner = VitestRunner()
-
-        message = """
-expect(received).toBe(expected)
-
-Expected: 2
-Received: 1
-        """
-
+        message = "expect(received).toBe(expected)\n\nExpected: 2\nReceived: 1"
         result = runner._extract_assertion(message)
         assert "Expected:" in result or "expect" in result.lower()
 
     def test_extract_expected_received_pattern(self) -> None:
-        """Test extracting Expected/Received pattern."""
         runner = VitestRunner()
-
-        message = """
-Expected: 42
-Received: 0
-        """
-
+        message = "\nExpected: 42\nReceived: 0\n"
         result = runner._extract_assertion(message)
         assert "Expected:" in result
 
     def test_extract_first_meaningful_line(self) -> None:
-        """Test extracting first meaningful line as fallback."""
         runner = VitestRunner()
-
-        message = """
-TypeError: Cannot read property 'foo' of undefined
-    at Object.<anonymous> (test.ts:5:1)
-        """
-
+        message = "\nTypeError: Cannot read property 'foo' of undefined\n    at Object.<anonymous> (test.ts:5:1)\n"
         result = runner._extract_assertion(message)
         assert "TypeError" in result
 
     def test_empty_message(self) -> None:
-        """Test empty message returns empty string."""
         runner = VitestRunner()
         result = runner._extract_assertion("")
         assert result == ""
 
     def test_extract_skips_short_lines(self) -> None:
-        """Test that short lines (<=5 chars) are skipped in fallback."""
         runner = VitestRunner()
-
-        message = """
-at Object.test
-Some meaningful error message here
-        """
-
+        message = "\nat Object.test\nSome meaningful error message here\n"
         result = runner._extract_assertion(message)
         assert len(result) > 5
 
 
 class TestVitestIssueIdGeneration:
-    """Tests for deterministic issue ID generation."""
+    """Tests for deterministic issue ID generation (via base class)."""
 
     def test_same_input_same_id(self) -> None:
-        """Test same input produces same ID."""
         runner = VitestRunner()
-
         id1 = runner._generate_issue_id("Test > should work", "expect")
         id2 = runner._generate_issue_id("Test > should work", "expect")
-
         assert id1 == id2
 
     def test_different_input_different_id(self) -> None:
-        """Test different input produces different ID."""
         runner = VitestRunner()
-
         id1 = runner._generate_issue_id("Test > should work", "expect 1")
         id2 = runner._generate_issue_id("Test > should fail", "expect 2")
-
         assert id1 != id2
 
     def test_id_format(self) -> None:
-        """Test ID format starts with vitest-."""
         runner = VitestRunner()
-
         issue_id = runner._generate_issue_id("Test > should work", "expect")
-
         assert issue_id.startswith("vitest-")
-        assert len(issue_id) == len("vitest-") + 12  # 12 char hash
+        assert len(issue_id) == len("vitest-") + 12
